backend/server/lib/files/requirements.txt
annotated-types==0.7.0
anyio==4.9.0
cachetools==5.5.2
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
cryptography==44.0.3
google-auth==2.39.0
google-genai==1.13.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
pdfminer.six==20250327
pdfplumber==0.11.6
pillow==11.2.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.11.4
pydantic_core==2.33.2
PyMuPDF==1.25.5
pypdf==5.4.0
pypdfium2==4.30.1
python-dotenv==1.1.0
requests==2.32.3
rsa==4.9.1
sniffio==1.3.1
typing-inspection==0.4.0
typing_extensions==4.13.2
urllib3==2.4.0
websockets==15.0.1
